remove compiler warnings and errors with gcc/g++ 8.3.0

+++ devices/rawfiles/rawfiles.cpp
@@ -27,6 +27,7 @@
 #include <fcntl.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <inttypes.h>
 #include <sys/time.h>
 #include <time.h>
 #include <unistd.h>
@@ -126,8 +127,8 @@ void rawFiles::run(void) {
   bool eofReached = false;
 
   running.store(true);
-  period = (32768 * 1000) / 2048;  // full IQs read
-  fprintf(stderr, "Period = %lld\n", period);
+  period = ((int64_t)32768 * 1000) / 2048;  // full IQs read
+  fprintf(stderr, "Period = %" PRId64 "\n", period);
   bi = new std::complex<float>[bufferSize];
   nextStop = getMyTime();
   while (running.load()) {
